Drop patch versions from dependencies in Cargo.toml
According to semantic versioning rules, we should be capable of
compiling with any crate that is at least minor-version compatible with
our requirements. Retaining the patch versions here creates a lot of
unnecessary work for Dependabot and requires the maintainer to manually
bump the versions through PRs. With this change, this will no longer be
necessary for patch version changes, which will be applied to
`tower-lsp` automatically where applicable by Cargo.