Return Err from poll_ready() if stopped
According to the [docs] for `tower_service::Service::poll_ready()`, the
method implementation is supposed to return `Poll::Ready(Err(_))` if the
service in question is no longer able to accept requests. In our current
implementation, we return `Poll::Pending`, which is semantically
incorrect.

[docs]: https://docs.rs/tower-service/0.3.0/tower_service/trait.Service.html#tymethod.poll_ready

This changes our `poll_ready()` implementation to conform with the
upstream interface specification so we can properly handle backpressure.