Use Ordering::Relaxed when incrementing client request ID
The relevant chapter of the [Nomicon] states that `Ordering::Relaxed`
should be safe to use for incrementing an atomic counter as long as the
datum isn't being used to communicate between threads operating on
shared memory.

[Nomicon]: https://doc.rust-lang.org/nomicon/atomics.html#relaxed

In our case, we are using it to generate unique request IDs, so this
should be safe and would provide a performance improvement on platforms
with weak ordering semantics.