Fix bidirectional request routing by using buffer_unordered()
It seems that the previous implementation of server-to-client request
routing was broken because of `service.call(request).await` in `Server`
blocking forever, preventing the response from being routed back to the
awaiting RPC handler.

This commit fixes this by rewriting `Server::serve() to use
`buffer_unordered()` to drive at least 4 concurrent futures at the
same time, thereby avoiding stalling the executor and enabling proper
multiplexing of client-to-server and server-to-client communication.

As a bonus, we also eliminate one use of `tokio::spawn()` by using
`futures::join!()` to drive the printer and reader futures concurrently
without depending on a threaded executor. This is a nice step towards
eventually decoupling from `tokio`.