Allow !Send and !'static streams and IO handles
This has been a long-standing issue resulting from the strict
requirements imposed by the `tokio::spawn()` call in `Server::serve()`.
But now that we rely on `join!()` to drive both futures to completion
without having to deal with the complexities of a multi-threaded
executor, we no longer require that the interleaved stream and
`AsyncRead`/`AsyncWrite` handles be `Send + 'static`.

This finally allows us to update the `stdio.rs` tests to take mock
`stdin` and `stdout` handles as `&mut`, allowing us to inspect the final
state once the test is done and the `Server` has shut down.