Expand lsp-types dependency with semver range
Because the changes to `lsp-types` from 0.74 to 0.76 don't require any
breaking changes to `tower-lsp`, we can affort to use a version range
for this dependency. If the next minor version comes out with breaking
changes that affect `tower-lsp`, we will ditch this version range for a
single version spec again.