Allow language servers to store Client as struct field
Currently, `Client` is only ever exposed to the user as an argument
explicitly passed by reference into some (but not all) `LanguageServer`
trait methods. This was originally done for the sake of implementation
simplicity and getting a quick MVP out the door, but it is rapidly
becoming clear that this design does not scale.

User input has shown that non-trivial language server designs would
greatly benefit from being passed an owned `Client` at initialization
time instead of being passed in at runtime through a callback.
Additionally, routing the `Client` through each individual callback is
expensive to maintain in the long run. The current design, however,
makes migrating away from the status quo impossible.

To achieve this new goal, we modify the `LspService::new()` constructor
to accept a closure of type `FnOnce(Client) -> T` instead of an
initialized `T` value directly. This allows the construction of the
server struct to be deferred until the `Client` has been initialized
internally. With this approach, `LanguageServer` implementers can be
initialized with a `Client` field in a type-safe way without requiring
the user to manually track the initialization state of `Client` in their
code. Additional resources required by the language server struct can be
initialized ahead of time in `main()` and passed into the closure, if
needed.