Define pending requests data structures for server/client
In a nutshell, both `ServerRequests` and `ClientRequests` allow us to
mark certain JSON-RPC request IDs as "pending" and uniquely associate
them with an asynchronous task. Though both data structures appear
similar on the surface, their internal designs and respective purposes
are slightly different.

`ClientRequests` is optimized for sending requests to the client and
routing responses back to the server. By contrast, `ServerRequests` is
geared towards receiving requests from the client and triggering the
corresponding `LanguageServer` async handlers using the `.execute()`
method. Also, pending requests stored in the `ServerRequests` map can be
canceled by the user, while those in `ClientRequests` cannot.

These two data structures are meant for internal use only and will not
be exposed to the user.