Reject multiple initialize requests sent in quick succession
It is possible for the `initialize` request handler to run more than
once if the client spams the server with multiple `initialize` requests
in quick succession without waiting for a response. This behavior is
unlikely to happen since it goes against the LSP spec and the client
would have to be hopelessly buggy, but we should cover this case
nonetheless.

To achieve this, we add a new `Initializing` state machine variant which
prevents further `initialize` requests from being accepted until the
current one has finished processing and responded back to the client.