Report missing params as "invalid params" instead of "parse error"
These changes let us identify when a required `params` field has been
omitted and correctly report it as a "invalid params" error instead of a
"parse error" and also provide detailed output from `serde_json::Error`
in the `message` field, if there is any.

We also take the opportunity to eliminate most of the noisy duplicate
logging calls in `handle_request()`. We will provide this info to the
user at the transport layer instead, starting in the next commit.