Support 'null' values as request IDs
Despite the Language Server Protocol specification claiming otherwise
[here], the `id` field of a request object _can_ be `null` according to
the JSON-RPC 2.0 specification:

[here]: https://microsoft.github.io/language-server-protocol/specification#requestMessage

Although the use of `null` as a request ID value is rare in practice and
explicitly discouraged by the JSON-RPC 2.0 specification, we must be
compatible with as many JSON-RPC client implementations as possible.
Therefore, we must support parsing `null` request IDs in incoming
messages, even if we never emit them ourselves.