Fix semver breakage with tower-lsp-macros crate
The `tower-lsp-macros` crate should have been released with a minor
version bump, not a patch version bump. This commit updates the versions
appropriately in the `Cargo.toml` manifests in preparation for yanking
and re-releasing the same code under the new designations `tower-lsp`
0.15.1 and `tower-lsp-macros` 0.5.0.

This is to address a semver incompatibility introduced in PR #285.