Support mutable methods and relax Send bounds
Now that we have a runtime-agnostic single-threaded async-aware
synchronization primitive, we can add support for mutable methods on the
`LanguageServer` trait and relax the `Send` bounds on all futures. This
approach is definitely going to be iterated on, and unit tests will need
to be added/updated later.

This commit also temporarily comments out a single unit test in
`src/service/pending.rs` due to a non-`Send` future not being spawnable
on the work-stealing `tokio` executor.